Fix MSVC x64 warnings C4267: conversion from 'size_t' to 'int', possible loss of data

+++ fmatvec/ast.cc
@@ -388,7 +388,7 @@ Symbol::Symbol(const boost::uuids::uuid& uuid_) : version(0), uuid(uuid_) {}
 string Symbol::getUUIDStr() const {
 #ifndef NDEBUG // FMATVEC_DEBUG_SYMBOLICEXPRESSION_UUID
   if(getenv("FMATVEC_DEBUG_SYMBOLICEXPRESSION_UUID")) {
-    auto res=mapUUIDInt.insert(make_pair(uuid, mapUUIDInt.size()+1));
+    auto res=mapUUIDInt.insert(make_pair(uuid, static_cast<int>(mapUUIDInt.size())+1));
     return "s"+to_string(res.first->second);
   }
   return to_string(uuid);
